docs: update comments
