+++ exact-numbers/src/test/kotlin/expressions/term/TermTest.kt
@@ -1,10 +1,8 @@
 package expressions.term
 
 import exactnumbers.exactfraction.ExactFraction
-import exactnumbers.irrationals.common.Irrational
 import exactnumbers.irrationals.log.Log
 import exactnumbers.irrationals.pi.Pi
-import java.math.BigDecimal
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
@@ -18,102 +16,14 @@ internal class TermTest {
     private val logNum4 = Log(ExactFraction(25, 121))
     private val one = ExactFraction.ONE
 
-    @Test
-    internal fun testConstructor() {
-        // zero
-        var expectedCoeff = ExactFraction.ZERO
-        var expectedNumbers: List<Irrational> = listOf()
+    @Test internal fun testConstructor() = runConstructorTests()
+    @Test internal fun testFromValues() = runFromValuesTests()
 
-        var term = Term(ExactFraction.ZERO, listOf())
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
+    @Test internal fun testTimes() = runTimesTests()
+    @Test internal fun testDiv() = runDivTests()
 
-        term = Term(ExactFraction.ZERO, listOf(logNum1, logNum3, Pi()))
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-
-        term = Term(ExactFraction.FIVE, listOf(logNum1, Log.ZERO))
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-
-        // others
-        term = Term(-ExactFraction.FIVE, listOf())
-        expectedCoeff = -ExactFraction.FIVE
-        expectedNumbers = listOf()
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-
-        term = Term(ExactFraction(17, 3), listOf(Pi(), logNum2))
-        expectedCoeff = ExactFraction(17, 3)
-        expectedNumbers = listOf(Pi(), logNum2)
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-
-        term = Term(
-            ExactFraction(-4, 5),
-            listOf(Pi(true), logNum2, logNum1, Pi(), logNum1.swapDivided(), logNum2)
-        )
-        expectedCoeff = ExactFraction(-4, 5)
-        expectedNumbers = listOf(Pi(true), logNum2, logNum1, Pi(), logNum1.swapDivided(), logNum2)
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-    }
-
-    @Test
-    internal fun testFromValues() {
-        // zero
-        var expectedCoeff = ExactFraction.ZERO
-        var expectedNumbers: List<Irrational> = listOf()
-
-        var term = Term.fromValues(ExactFraction.ZERO, listOf(), 0)
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-
-        term = Term.fromValues(ExactFraction.ZERO, listOf(logNum1, Log.ONE), 8)
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-
-        term = Term.fromValues(-ExactFraction.EIGHT, listOf(logNum1, logNum2, Log.ZERO, logNum3), 5)
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-
-        // nonzero
-        term = Term.fromValues(ExactFraction(-5, 7), listOf(), 0)
-        expectedCoeff = ExactFraction(-5, 7)
-        expectedNumbers = listOf()
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-
-        term = Term.fromValues(ExactFraction.EIGHT, listOf(), 3)
-        expectedCoeff = ExactFraction.EIGHT
-        expectedNumbers = listOf(Pi(), Pi(), Pi())
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-
-        term = Term.fromValues(ExactFraction.EIGHT, listOf(), -3)
-        expectedCoeff = ExactFraction.EIGHT
-        expectedNumbers = listOf(Pi(true), Pi(true), Pi(true))
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-
-        term = Term.fromValues(ExactFraction(-2, 191), listOf(logNum1, logNum2, logNum1), 0)
-        expectedCoeff = ExactFraction(-2, 191)
-        expectedNumbers = listOf(logNum1, logNum2, logNum1)
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-
-        term = Term.fromValues(ExactFraction(-2, 191), listOf(logNum1, logNum2, logNum3), 2)
-        expectedCoeff = ExactFraction(-2, 191)
-        expectedNumbers = listOf(logNum1, logNum2, logNum3, Pi(), Pi())
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-
-        term = Term.fromValues(ExactFraction(22), listOf(logNum1, logNum2, logNum3), -2)
-        expectedCoeff = ExactFraction(22)
-        expectedNumbers = listOf(logNum1, logNum2, logNum3, Pi(true), Pi(true))
-        assertEquals(expectedCoeff, term.coefficient)
-        assertEquals(expectedNumbers, term.numbers)
-    }
+    @Test internal fun testGetSimplified() = runGetSimplifiedTests()
+    @Test internal fun testGetValue() = runGetValueTests()
 
     @Test
     internal fun testEquals() {
@@ -265,129 +175,6 @@ internal class TermTest {
     }
 
     @Test
-    internal fun testGetSimplified() {
-        // simplified
-        var term = Term(ExactFraction.EIGHT, listOf(Pi(), Pi(true)))
-        var expectedCoeff = ExactFraction.EIGHT
-        var expectedNumbers: List<Irrational> = listOf()
-        var result = term.getSimplified()
-        assertEquals(expectedCoeff, result.coefficient)
-        assertEquals(expectedNumbers, result.numbers)
-
-        term = Term(ExactFraction(-3, 2), listOf(Pi(), Pi(true), Pi(), logNum1))
-        expectedCoeff = ExactFraction(-3, 2)
-        expectedNumbers = listOf(logNum1, Pi())
-        result = term.getSimplified()
-        assertEquals(expectedCoeff, result.coefficient)
-        assertEquals(expectedNumbers, result.numbers)
-
-        term = Term(ExactFraction.HALF, listOf(Log.ONE, logNum1))
-        expectedCoeff = ExactFraction.HALF
-        expectedNumbers = listOf(logNum1)
-        result = term.getSimplified()
-        assertEquals(expectedCoeff, result.coefficient)
-        assertEquals(expectedNumbers, result.numbers)
-
-        term = Term(-ExactFraction.HALF, listOf(Log.ONE, Pi(true)))
-        expectedCoeff = -ExactFraction.HALF
-        expectedNumbers = listOf(Pi(true))
-        result = term.getSimplified()
-        assertEquals(expectedCoeff, result.coefficient)
-        assertEquals(expectedNumbers, result.numbers)
-
-        term = Term(ExactFraction(18, 5), listOf(Pi(true), logNum2, Pi(true), logNum2, logNum1, logNum2.swapDivided(), Pi(true), Pi()))
-        expectedCoeff = ExactFraction(18, 5)
-        expectedNumbers = listOf(logNum1, logNum2, Pi(true), Pi(true))
-        result = term.getSimplified()
-        assertEquals(expectedCoeff, result.coefficient)
-        assertEquals(expectedNumbers, result.numbers)
-
-        // no changes
-        term = Term(ExactFraction.EIGHT, listOf())
-        expectedCoeff = ExactFraction.EIGHT
-        expectedNumbers = listOf()
-        result = term.getSimplified()
-        assertEquals(expectedCoeff, result.coefficient)
-        assertEquals(expectedNumbers, result.numbers)
-
-        term = Term(ExactFraction.EIGHT, listOf(logNum1))
-        expectedCoeff = ExactFraction.EIGHT
-        expectedNumbers = listOf(logNum1)
-        result = term.getSimplified()
-        assertEquals(expectedCoeff, result.coefficient)
-        assertEquals(expectedNumbers, result.numbers)
-
-        term = Term(ExactFraction(-5, 6), listOf(Pi(true)))
-        expectedCoeff = ExactFraction(-5, 6)
-        expectedNumbers = listOf(Pi(true))
-        result = term.getSimplified()
-        assertEquals(expectedCoeff, result.coefficient)
-        assertEquals(expectedNumbers, result.numbers)
-
-        term = Term(ExactFraction.SEVEN, listOf(Pi(), Pi(), logNum1, logNum4, logNum2.swapDivided()))
-        expectedCoeff = ExactFraction.SEVEN
-        expectedNumbers = listOf(logNum4, logNum1, logNum2.swapDivided(), Pi(), Pi())
-        result = term.getSimplified()
-        assertEquals(expectedCoeff, result.coefficient)
-        assertEquals(expectedNumbers, result.numbers)
-    }
-
-    @Test
-    internal fun testGetValue() {
-        // just number
-        var term = Term.ZERO
-        var expected = BigDecimal.ZERO
-        assertEquals(expected, term.getValue())
-
-        term = Term(ExactFraction.EIGHT, listOf())
-        expected = BigDecimal("8")
-        assertEquals(expected, term.getValue())
-
-        term = Term(ExactFraction(-1, 3), listOf())
-        expected = BigDecimal("-0.33333333333333333333")
-        assertEquals(expected, term.getValue())
-
-        // just logs
-        term = Term(one, listOf(logNum1, logNum1.swapDivided()))
-        expected = BigDecimal.ONE
-        assertEquals(expected, term.getValue())
-
-        term = Term(one, listOf(Log(ExactFraction(3333))))
-        expected = BigDecimal("3.52283531366053")
-        assertEquals(expected, term.getValue())
-
-        term = Term(one, listOf(logNum1, logNum2))
-        expected = BigDecimal("0.61342218956802803344500481172832")
-        assertEquals(expected, term.getValue())
-
-        // just pi
-        term = Term(one, listOf(Pi(), Pi(true)))
-        expected = BigDecimal.ONE
-        assertEquals(expected, term.getValue())
-
-        term = Term(one, listOf(Pi(true)))
-        expected = BigDecimal("0.31830988618379069570")
-        assertEquals(expected, term.getValue())
-
-        term = Term(one, listOf(Pi(), Pi(), Pi()))
-        expected = BigDecimal("31.006276680299813114880451174049119330924860257")
-        assertEquals(expected, term.getValue())
-
-        // combination
-        term = Term(ExactFraction(-8, 3), listOf(Pi()))
-        expected = BigDecimal("-8.3775804095727813333")
-        assertEquals(expected, term.getValue())
-
-        term = Term(ExactFraction.SEVEN, listOf(Log(ExactFraction(6, 7), 4), Pi(), Log(ExactFraction(40))))
-        expected = BigDecimal("-3.9175691871908989550925271506970548502975209152")
-        assertEquals(expected, term.getValue())
-
-        term = Term(ExactFraction.HALF, listOf(Log(ExactFraction.FOUR, 2, true), Pi(true), Log(ExactFraction(123456789))))
-        expected = BigDecimal("0.64390230285929702583103243749028475")
-        assertEquals(expected, term.getValue())
-    }
-
-    @Test
     internal fun testGetLogs() {
         // empty
         var expected: List<Log> = listOf()
@@ -515,7 +302,4 @@ internal class TermTest {
         expected = "<[-100/333]x${Pi(true)}x${logNum2}x${logNum2}x${logNum4}x${Pi()}x$logNum1>"
         assertEquals(expected, term.toString())
     }
-
-    @Test internal fun testTimes() = runTimesTests()
-    @Test internal fun testDiv() = runDivTests()
 }
